Updated message type type

+++ BroadCastTransfer.cpp
@@ -3,7 +3,7 @@
 BroadcastTransfer::BroadcastTransfer(
     uint8_t *payload,
     uint16_t payloadLength,
-    int dataTypeId,
+    uint16_t dataTypeId,
     uint8_t sourceNodeId,
     uint8_t priority,
     uint8_t transferId)
