+++ TransferBase.cpp
@@ -3,7 +3,7 @@
 TransferBase::TransferBase(
     uint8_t *payload,
     uint16_t payloadLength,
-    int dataTypeId,
+    uint16_t dataTypeId,
     uint8_t priority,
     uint8_t transferId)
 {
@@ -16,6 +16,6 @@ TransferBase::TransferBase(
 
 uint8_t *TransferBase::getPayload() { return _payload; }
 uint16_t TransferBase::getPayloadLength() { return _payloadLength; }
-int TransferBase::getDataTypeId() { return _dataTypeId; }
+uint16_t TransferBase::getDataTypeId() { return _dataTypeId; }
 uint8_t TransferBase::getPriority() { return _priority; }
 uint8_t TransferBase::getTransferId() { return _transferId; }
