+++ TransferBase.h
@@ -4,17 +4,17 @@
 class TransferBase
 {
 public:
-  TransferBase(uint8_t *, uint16_t, int, uint8_t, uint8_t);
+  TransferBase(uint8_t *, uint16_t, uint16_t, uint8_t, uint8_t);
   uint8_t *getPayload();
   uint16_t getPayloadLength();
-  int getDataTypeId();
+  uint16_t getDataTypeId();
   uint8_t getPriority();
   uint8_t getTransferId();
 
 private:
   uint8_t *_payload;
   uint16_t _payloadLength;
-  int _dataTypeId;
+  uint16_t _dataTypeId;
   uint8_t _priority;
   uint8_t _transferId;
 };
